fixed build error in sales-point page
