+++ public/sw.js
@@ -1,92 +1,1 @@
-/**
- * Copyright 2018 Google Inc. All Rights Reserved.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// If the loader is already loaded, just stop.
-if (!self.define) {
-  let registry = {};
-
-  // Used for `eval` and `importScripts` where we can't get script URL by other means.
-  // In both cases, it's safe to use a global var because those functions are synchronous.
-  let nextDefineUri;
-
-  const singleRequire = (uri, parentUri) => {
-    uri = new URL(uri + ".js", parentUri).href;
-    return registry[uri] || (
-      
-        new Promise(resolve => {
-          if ("document" in self) {
-            const script = document.createElement("script");
-            script.src = uri;
-            script.onload = resolve;
-            document.head.appendChild(script);
-          } else {
-            nextDefineUri = uri;
-            importScripts(uri);
-            resolve();
-          }
-        })
-      
-      .then(() => {
-        let promise = registry[uri];
-        if (!promise) {
-          throw new Error(`Module ${uri} didn’t register its module`);
-        }
-        return promise;
-      })
-    );
-  };
-
-  self.define = (depsNames, factory) => {
-    const uri = nextDefineUri || ("document" in self ? document.currentScript.src : "") || location.href;
-    if (registry[uri]) {
-      // Module is already loading or loaded.
-      return;
-    }
-    let exports = {};
-    const require = depUri => singleRequire(depUri, uri);
-    const specialDeps = {
-      module: { uri },
-      exports,
-      require
-    };
-    registry[uri] = Promise.all(depsNames.map(
-      depName => specialDeps[depName] || require(depName)
-    )).then(deps => {
-      factory(...deps);
-      return exports;
-    });
-  };
-}
-define(['./workbox-631a4576'], (function (workbox) { 'use strict';
-
-  importScripts();
-  self.skipWaiting();
-  workbox.clientsClaim();
-  workbox.registerRoute("/", new workbox.NetworkFirst({
-    "cacheName": "start-url",
-    plugins: [{
-      cacheWillUpdate: async ({
-        response: e
-      }) => e && "opaqueredirect" === e.type ? new Response(e.body, {
-        status: 200,
-        statusText: "OK",
-        headers: e.headers
-      }) : e
-    }]
-  }), 'GET');
-  workbox.registerRoute(/.*/i, new workbox.NetworkOnly({
-    "cacheName": "dev",
-    plugins: []
-  }), 'GET');
-
-}));
+if(!self.define){let e,s={};const a=(a,i)=>(a=new URL(a+".js",i).href,s[a]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()})).then((()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(i,n)=>{const c=e||("document"in self?document.currentScript.src:"")||location.href;if(s[c])return;let t={};const r=e=>a(e,c),u={module:{uri:c},exports:t,require:r};s[c]=Promise.all(i.map((e=>u[e]||r(e)))).then((e=>(n(...e),t)))}}define(["./workbox-c2c0676f"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/dynamic-css-manifest.json",revision:"d751713988987e9331980363e24189ce"},{url:"/_next/static/chunks/173-7ddf08a60e28afe0.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/203.2b4c1ee4fbe3a7cf.js",revision:"2b4c1ee4fbe3a7cf"},{url:"/_next/static/chunks/218.57a830a2c55ba802.js",revision:"57a830a2c55ba802"},{url:"/_next/static/chunks/277-975124e7237ef95c.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/447-0f5eaca39a5016d2.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/4bd1b696-5380540ad3a6d97b.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/738-6bb01dfd63c8b776.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/863-e60a8895de297cfe.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/868-d6058286cf3d53da.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/_not-found/page-d7aa5a62d7aa0148.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/layout-bd6a3abc561fd496.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/auth/authorize-device/page-fbcb5e08314e5311.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/auth/forgot-password/page-ea2337f4c40cc8b2.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/auth/login-attempt-notification/page-4e3f70e408a3d2d1.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/auth/login/page-b4e32f95cb86d492.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/auth/login/sales-point/login-new/page-2665df2821e73b6e.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/auth/login/sales-point/page-2d0e4371bb00d82e.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/auth/reset-password/page-0ade70e23983a0da.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/auth/sign-up/page-f6f72a503748768c.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/auth/verify-otp/page-ff972ce37f6c4444.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/dashboard/admin/page-f905cf9ad2157295.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/dashboard/manager/page-0805ad43ff8f1c73.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/dashboard/sales-point/page-26ca5dd879a7e350.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/splash/splash1/page-1935d77943f40d3c.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/splash/splash2/page-cc0469ffa0120731.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/app/pages/splash/splash3/page-b06ad28e24ced6d3.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/framework-6b27c2b7aa38af2d.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/main-0d164eb93b185bbe.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/main-app-dd90453daeab121e.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/pages/_app-430fec730128923e.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/pages/_error-2d7241423c4a35ba.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-e5c0edd1aaf3c2e4.js",revision:"eGF-0ZiFBQEBisuguSQmp"},{url:"/_next/static/css/69cc9dc15dfbf8b9.css",revision:"69cc9dc15dfbf8b9"},{url:"/_next/static/css/951161188d05844f.css",revision:"951161188d05844f"},{url:"/_next/static/css/e7eddfe4828c1690.css",revision:"e7eddfe4828c1690"},{url:"/_next/static/eGF-0ZiFBQEBisuguSQmp/_buildManifest.js",revision:"6df351e08c63ea2c6b9f3938ebf38c40"},{url:"/_next/static/eGF-0ZiFBQEBisuguSQmp/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/assets/Danger.png",revision:"05d405dee42b677ffbadd169e851efc8"},{url:"/assets/Shield.png",revision:"ca3eb0c0dbfafb9964955530504b9bc4"},{url:"/assets/Star.png",revision:"46010e6c54484b8d826758094d7e18f5"},{url:"/assets/anonymous.png",revision:"b48f5bab742bd075bef3bd1e281bdfee"},{url:"/assets/arrow_down.png",revision:"776aeef4e8d7d3190c62712b193e533c"},{url:"/assets/arrow_left.png",revision:"7e48f3f91faa795f882911fdd299e995"},{url:"/assets/arrow_right.png",revision:"7941b0551dbf50e9773419493180c44d"},{url:"/assets/back.png",revision:"7e48f3f91faa795f882911fdd299e995"},{url:"/assets/back_green.png",revision:"d05ca06e56bae599af93a698871db85c"},{url:"/assets/brand_mark1.png",revision:"3d0b45e9651f5a4bcd48b797539e0df0"},{url:"/assets/brand_mark2.png",revision:"8e160ad6cc11a3d974f0dba45b3649ce"},{url:"/assets/chatbox.png",revision:"cc9cdaff5364dc43c696ef61308b217d"},{url:"/assets/checkbox_gray.png",revision:"2ac4b99696947ac6d4846818f4131c86"},{url:"/assets/checkbox_green.png",revision:"82039f52f7c40483594d61918d532683"},{url:"/assets/dashboard.png",revision:"f0adc8a6f54efe8be0cd8803a0a3acce"},{url:"/assets/dashboard_active.png",revision:"3a6d4b233869d54b794af5ad96b1413f"},{url:"/assets/error.png",revision:"00f75c29676e2dd3e0885fb1029d5095"},{url:"/assets/eye-closed-green.png",revision:"cb7b62a0e0f035135616e9406797aa53"},{url:"/assets/eye-closed.png",revision:"58efb3b76ad3ac3bd0b51455cf8c91a9"},{url:"/assets/eye-open-green.png",revision:"dd7bfb5ff0883ccd73e28abd927afb6a"},{url:"/assets/eye-open.png",revision:"59ec4cf988cd36b88706e6116505be93"},{url:"/assets/logo_blue.png",revision:"ab2cad31058c7256fe70263d89f1d7df"},{url:"/assets/logo_long.png",revision:"a58ba0cd4730a1ad9e8badf9334f6e34"},{url:"/assets/logo_white.png",revision:"78f22d7f54acca99269478b68003ba1e"},{url:"/assets/logout.png",revision:"e5c4c36b9f5341797567e9ea44d1613f"},{url:"/assets/material.png",revision:"ae16091bf89850380139ead0b6487db1"},{url:"/assets/material_active.png",revision:"cbea261eca3adc15a9c044fa504dbdda"},{url:"/assets/notification.png",revision:"2a83101074b0c8a02d2f38cb78e1b7e8"},{url:"/assets/picture1.jpg",revision:"551e3c2414874171e66b7bee3d730f8d"},{url:"/assets/picture2.jpg",revision:"a0af4a1ea0665959a64414d030cdd1ca"},{url:"/assets/picture3.jpg",revision:"e8e914e171193ba9a3f27b7a545cf2b3"},{url:"/assets/picture4.jpg",revision:"111279c90fc50a6d2106cf1f549ba002"},{url:"/assets/picture5.jpg",revision:"b167122fc929a4f3ba980e8bc3bc2d4e"},{url:"/assets/product.png",revision:"a6bd6ef2d852d954c085f4bc70015dda"},{url:"/assets/product_active.png",revision:"53fff18da209eaba0ccdcaa5bab3d035"},{url:"/assets/profile.png",revision:"992f55cbc72f6f96759bf6813451f3dc"},{url:"/assets/profile_active.png",revision:"305a40f78f005ea5279776c2085592de"},{url:"/assets/report_active.png",revision:"65bff1d9fa9440af4ed32f772c62389a"},{url:"/assets/report_white.png",revision:"62184919779ad5169037deadb05f1511"},{url:"/assets/settings.png",revision:"5caa20173a24e9acde0b77bd9b00b09b"},{url:"/assets/settings_active.png",revision:"e0112d4c219a57ecb7ce6db8e5325589"},{url:"/assets/store.png",revision:"300483a4ff7e17dad02ca5f2bbd6ea08"},{url:"/assets/store_active.png",revision:"de061c05a8146146b7333f5b6f26feee"},{url:"/assets/toggle-off.png",revision:"55d388b4fe451a2f1e603a2048813ae6"},{url:"/assets/verified.png",revision:"80374c513160919cb7d579d89d7135d1"},{url:"/assets/warning.png",revision:"afb7361ec685f946f9d918de575008f5"},{url:"/icon512_maskable.png",revision:"1c1abc12a0bda1cbbaa2a074fb3ed2a0"},{url:"/icon512_rounded.png",revision:"7ddda3fa29cd48ada0e4c838a26ddb52"},{url:"/images/welcome.png",revision:"009761afc3a36c62bace1ca68e246b90"},{url:"/manifest.json",revision:"51ff3c5c1d558295570cad24c524089d"}],{ignoreURLParametersMatching:[/^utm_/,/^fbclid$/]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({response:e})=>e&&"opaqueredirect"===e.type?new Response(e.body,{status:200,statusText:"OK",headers:e.headers}):e}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:2592e3})]}),"GET"),e.registerRoute(/\/_next\/static.+\.js$/i,new e.CacheFirst({cacheName:"next-static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4|webm)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:48,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({sameOrigin:e,url:{pathname:s}})=>!(!e||s.startsWith("/api/auth/callback")||!s.startsWith("/api/"))),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({request:e,url:{pathname:s},sameOrigin:a})=>"1"===e.headers.get("RSC")&&"1"===e.headers.get("Next-Router-Prefetch")&&a&&!s.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages-rsc-prefetch",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({request:e,url:{pathname:s},sameOrigin:a})=>"1"===e.headers.get("RSC")&&a&&!s.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages-rsc",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:{pathname:e},sameOrigin:s})=>s&&!e.startsWith("/api/")),new e.NetworkFirst({cacheName:"pages",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({sameOrigin:e})=>!e),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
